+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.733.0",
+    "@aws-sdk/s3-request-presigner": "^3.734.0",
     "@clerk/nextjs": "^6.9.15",
     "@radix-ui/react-dialog": "^1.1.4",
     "@radix-ui/react-label": "^2.1.1",
